[ADD] #20 - Moved render_image to /rendering instead of /raycasting. Removed .vscode. Moved Bonus functions
